Taggable interface approach + keeping inline methods PoC
